+++ demo/CypressLayout.tsx
@@ -23,6 +23,7 @@ import {
   MultiSelect,
   ToolTip,
   TimePicker,
+  Tour,
   Validation
 } from './cypress';
 import { Styles } from './components/Styles';
@@ -90,6 +91,9 @@ export const CypressLayout = () => (
         </Route>
         <Route path="/cypress/timepicker">
           <TimePicker />
+        </Route>
+        <Route path="/cypress/tour">
+          <Tour />
         </Route>
         <Route path="/cypress/validation">
           <Validation />