+++ leda/components/Tour/helpers.ts
@@ -1,10 +1,45 @@
 import * as React from 'react';
 
+/**
+ * Helper creates point
+ * @param {number} x - x coordinate of point
+ * @param {number} y - y coordinate of point
+ *
+ * @returns { {x: number; y: number;} } - coordinates of point
+ */
 const createPoint = (x: number, y: number) => ({
   x, y,
 });
 
-export const createOverlaySvgPath = (element: HTMLElement | null, borderRadius: number): string => {
+/**
+ * Helper sets default tour element styles
+ * @param {HTMLElement | null | undefined} element - tour element
+ */
+export const setElementDefaultStyles = (element: HTMLElement | null | undefined): void => {
+  if (!element) return;
+  element.style.position = '';
+  element.style.zIndex = '';
+};
+
+/**
+ * Helper sets tour element styles
+ * @param {HTMLElement | null | undefined} element - tour element
+ */
+export const setElementStyles = (element: HTMLElement | null | undefined): void => {
+  if (!element) return;
+  element.style.position = 'relative';
+  element.style.zIndex = '10002';
+};
+
+/**
+ * Helper creates path of svg overlay element
+ * @param {HTMLElement | null} element - tour element
+ * @param {number} borderRadius - border radius of highlighted space
+ * @param {number} padding - padding of highlighted space
+ *
+ * @returns {string} - path of svg overlay element
+ */
+export const createOverlaySvgPath = (element: HTMLElement | null, borderRadius: number, padding: number): string => {
   const windowWidth = window.innerWidth;
   const windowHeight = window.innerHeight;
 
@@ -21,10 +56,10 @@ export const createOverlaySvgPath = (element: HTMLElement | null, borderRadius: 
 
   const rect = element.getBoundingClientRect();
 
-  const elementTopLeft = createPoint(rect.left, rect.top);
-  const elementTopRight = createPoint(rect.right, rect.top);
-  const elementBottomRight = createPoint(rect.right, rect.bottom);
-  const elementBottomLeft = createPoint(rect.left, rect.bottom);
+  const elementTopLeft = createPoint(rect.left - padding, rect.top - padding);
+  const elementTopRight = createPoint(rect.right + padding, rect.top - padding);
+  const elementBottomRight = createPoint(rect.right + padding, rect.bottom + padding);
+  const elementBottomLeft = createPoint(rect.left - padding, rect.bottom + padding);
 
   return `
     M 0 0
@@ -45,6 +80,14 @@ export const createOverlaySvgPath = (element: HTMLElement | null, borderRadius: 
     Z`;
 };
 
+/**
+ * Helper calculates modal window position
+ * @param {string} position - position of modal window against tour element
+ * @param {HTMLElement} element - tour element
+ * @param {boolean} isScrolling - flag defines if document is scrolling
+ *
+ * @returns {React.CSSProperties} - styles of modal window
+ */
 export const getModalPositionStyles = (position: string, element: HTMLElement, isScrolling: boolean): React.CSSProperties => {
   if (isScrolling) {
     return {