+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Set up automatic calculation of test coverage, added aliases to demo folder
 - Added aliases for imports
+- Tour: Added padding prop and possibility to interact with tour element
 
 ### Fixed
 - DatePicker, TimePicker: Fixed validation